% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracies_Agg.R
\name{Accuracies_Agg}
\alias{Accuracies_Agg}
\title{Accuracies_Agg}
\usage{
Accuracies_Agg(listProphet_Results, interval_agg = "day",
  results_type = "test_set")
}
\arguments{
\item{listProphet_Results}{A list with outputs from ProphetWrapper function}

\item{interval_agg}{A character value defining the level to be used to aggregate the series to compute the error metrics. It should be either the same interval of the series or higher.}

\item{results_type}{A character value defining how to calculate the accuracy metrics (test_set is default).}
}
\value{
This function returns a list with 2 elements:
 \itemize{
 \item{\strong{Accuracies}}  {A data-frame with the accuracies of the best model for each element parsed in listProphet_Results argument.}
 \item{\strong{Graphs_Accuracy_Agg}} {A list with plots of actuals vs forecasts of the best model for each element parsed in listProphet_Results argument.}
}
}
\description{
Function used to explore the accuracy (MAPE, MSE, MAE, RMSE, MPE) for several models simultaneously, choosing the aggregation level
}
