% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_wrapper.R
\name{cv_wrapper}
\alias{cv_wrapper}
\title{cv_wrapper}
\usage{
cv_wrapper(df, period_param, horizon_param, model_param)
}
\arguments{
\item{df}{The full data set}

\item{period_param}{Size of the total length of periods tested/validated.}

\item{horizon_param}{The length of each of the cross-validation folds.}

\item{model_param}{The model}
}
\description{
Function used to simplify the cross validation workflow in ProphetWrapper
}
