% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultArguments.R
\name{DefaultArguments}
\alias{DefaultArguments}
\title{DefaultArguments}
\usage{
DefaultArguments(list_params, debug)
}
\arguments{
\item{list_params}{list of parameters}

\item{debug}{debbuger}
}
\description{
Function used to default parameters and generate the grid of parameters
}
