% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FutureRegressors.R
\name{FutureRegressors}
\alias{FutureRegressors}
\title{FutureRegressors:}
\usage{
FutureRegressors(final_forecasts_list_future_regressors,
  best_regressor1_future_regressors, best_regressor2_future_regressors,
  all_data_future_regressors, debug_future_regressors)
}
\arguments{
\item{final_forecasts_list_future_regressors}{List with either ProphetWrapper objects or future regressors and Date}

\item{best_regressor1_future_regressors}{Name of the best regressor 1}

\item{best_regressor2_future_regressors}{Name of the best regressor 2}

\item{all_data_future_regressors}{All data available}

\item{debug_future_regressors}{Debug}
}
\description{
This is an internal function which unique purpose is to prepare future regressors (not exported).
}
