% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateViz.R
\name{GenerateViz}
\alias{GenerateViz}
\title{GenerateViz}
\usage{
GenerateViz(df_best_model_viz, Final_Forecasts_viz, list_params_viz,
  plotFrom_viz, debug_viz)
}
\arguments{
\item{df_best_model_viz}{Best model post testing}

\item{Final_Forecasts_viz}{final future forecasts}

\item{list_params_viz}{list of parameters}

\item{plotFrom_viz}{plotFrom variable}

\item{debug_viz}{debug}
}
\description{
Function to generate visualisations:
}
