% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_prophet_function.R
\name{modelling_prophet_function}
\alias{modelling_prophet_function}
\title{modelling_prophet_function}
\usage{
modelling_prophet_function(df_all_modelling, df_test_modelling,
  df_train_modelling, modelling_type, list_params_modelling,
  changepoint.prior.scale.modelling, holidays.prior.scale.modelling,
  regressor.prior.scale.modelling, seasonality.prior.scale.modelling,
  regressor1.modelling, regressor2.modelling, judgmental_forecasts.modelling,
  holidays_modelling, final_predictions_modelling, debug_modelling)
}
\arguments{
\item{df_all_modelling}{All data}

\item{df_test_modelling}{test data-frame}

\item{df_train_modelling}{train data-frame}

\item{modelling_type}{Either 'test' or 'final'. This parameter controls whether we are running test/train forecasts or a final one.}

\item{list_params_modelling}{A list with parameters passed by the user}

\item{changepoint.prior.scale.modelling}{parameters from list.params}

\item{holidays.prior.scale.modelling}{parameters from list.params}

\item{regressor.prior.scale.modelling}{parameters from list.params}

\item{seasonality.prior.scale.modelling}{parameters from list.params}

\item{regressor1.modelling}{parameters from list.params}

\item{regressor2.modelling}{parameters from list.params}

\item{judgmental_forecasts.modelling}{judgmental forecast named vector}

\item{holidays_modelling}{holidays}

\item{final_predictions_modelling}{length of final predictions}

\item{debug_modelling}{to debug}
}
\description{
Function to run the Prophet models (not exported)
}
