% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\docType{data}
\name{sessions}
\alias{sessions}
\title{Volume of Website Sessions per day for Company Fake (imaginary company as you can imagine !)}
\format{An object of class tbl_df, tbl, data.frame.}
\usage{
sessions
}
\description{
This dataset contains time-series data of volume of sessions per day for the imaginary company Fake.
}
\keyword{datasets}
