% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateArguments.R
\name{ValidateArguments}
\alias{ValidateArguments}
\title{ValidateArguments}
\usage{
ValidateArguments(df, list_params, best_model_in, plotFrom,
  main_accuracy_metric, train_set_imp_perc, judgmental_forecasts,
  final_predictions, debug = FALSE)
}
\arguments{
\item{df}{A data-frame original}

\item{list_params}{list of params}

\item{best_model_in}{best_model_in}

\item{plotFrom}{plotFrom}

\item{main_accuracy_metric}{main_accuracy_metric}

\item{train_set_imp_perc}{train_set_imp_perc}

\item{judgmental_forecasts}{judgmental_forecasts}

\item{final_predictions}{final_predictions}

\item{debug}{debugger}
}
\description{
Function used to validate the arguments parsed to ProphetWrapper:
}
